+++ packages/backend-proxy-middleware/src/middleware.ts
@@ -47,7 +47,7 @@ const backendProxyMiddleware = async ({
     });
 
     await initI18n();
-    dotenv.config();
+    dotenv.config({ quiet: true });
     const router = express.Router();
 
     const configOptions = options.configuration?.options ?? {};