+++ packages/ui5-proxy-middleware/src/ui5/middleware.ts
@@ -64,7 +64,7 @@ const ui5ProxyMiddleware = async ({
         transports: [new UI5ToolingTransport({ moduleName: 'ui5-proxy-middleware' })]
     });
 
-    dotenv.config();
+    dotenv.config({ quiet: true });
     const config: UI5ProxyConfig = {
         ui5: {
             path: ['/resources', '/test-resources'],