+++ packages/ui5-application-inquirer/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @sap-ux/ui5-application-inquirer
 
+## 0.17.16
+
+### Patch Changes
+
+- Updated dependencies [03d3ea1]
+    - @sap-ux/project-access@1.35.21
+    - @sap-ux/inquirer-common@0.11.40
+    - @sap-ux/project-input-validator@0.6.77
+
+## 0.17.15
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.39
+
+## 0.17.14
+
+### Patch Changes
+
+- Updated dependencies [9a48e63]
+    - @sap-ux/ui5-info@0.13.20
+    - @sap-ux/inquirer-common@0.11.38
+
+## 0.17.13
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.37
+
 ## 0.17.12
 
 ### Patch Changes