+++ packages/ui5-library-inquirer/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @sap-ux/ui5-library-inquirer
 
+## 0.6.99
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.40
+- @sap-ux/project-input-validator@0.6.77
+
+## 0.6.98
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.39
+
+## 0.6.97
+
+### Patch Changes
+
+- Updated dependencies [9a48e63]
+    - @sap-ux/ui5-info@0.13.20
+    - @sap-ux/inquirer-common@0.11.38
+
+## 0.6.96
+
+### Patch Changes
+
+- @sap-ux/inquirer-common@0.11.37
+
 ## 0.6.95
 
 ### Patch Changes